ml-service/TJMoneyML/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
pandas
numpy
matplotlib
scikit-learn

# REST API
djangorestframework>=3.14.0
djoser>=2.2.0